pyproject.toml
[project]
name = "web-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-daisy>=1.1.0",
    "django-model-utils>=5.0.0",
    "django-taggit>=6.1.0",
    "markdown>=3.8.2",
    "psycopg>=3.2.9",
    "pygments>=2.19.2",
    "pymdown-extensions>=10.16.1",
    "python-decouple>=3.8",
]
